sql population

+++ templates/lists/list_to_do_lists.php
@@ -21,8 +21,12 @@
             <?php }
               else{ ?>
               <td class="item-uncheck" id="item<?=$item['idItem']?>"><?=$item['info']?></td>
-            <?php } ?>
+            <?php }
+            if(isset($_SESSION['username']) && $_SESSION['username'] != ''){
+              if(ListBelongsUser($dbh,$_SESSION['username'],$todolist['idList'])) {?>
               <td class="deleteItem" id="delete<?=$item['idItem']?>">X</td>
+              <?}
+            }?>
           </tr>
         <?php }
           if(isset($_SESSION['username']) && $_SESSION['username'] != ''){
